+++ src/test/kotlin/db/tables/PollMessagesTest.kt
@@ -3,7 +3,7 @@ package com.neelkamath.omniChatBackend.db.tables
 import com.neelkamath.omniChatBackend.DbExtension
 import com.neelkamath.omniChatBackend.createVerifiedUsers
 import com.neelkamath.omniChatBackend.db.*
-import com.neelkamath.omniChatBackend.graphql.dataTransferObjects.UpdatedMessage
+import com.neelkamath.omniChatBackend.graphql.dataTransferObjects.UpdatedPollMessage
 import com.neelkamath.omniChatBackend.graphql.routing.GroupChatPublicity
 import com.neelkamath.omniChatBackend.graphql.routing.MessageText
 import com.neelkamath.omniChatBackend.graphql.routing.PollInput
@@ -31,7 +31,7 @@ class PollMessagesTest {
                     .map { messagesNotifier.subscribe(UserId(it)).flowable.subscribeWith(TestSubscriber()) }
                 PollMessages.setVote(adminId, messageId, option1, vote = true)
                 awaitBrokering()
-                val actual = adminSubscriber.values().map { (it as UpdatedMessage).getMessageId() }
+                val actual = adminSubscriber.values().map { (it as UpdatedPollMessage).getMessageId() }
                 assertEquals(listOf(messageId), actual)
                 nonParticipantSubscriber.assertNoValues()
             }
@@ -48,7 +48,7 @@ class PollMessagesTest {
             val subscriber = chatMessagesNotifier.subscribe(ChatId(chatId)).flowable.subscribeWith(TestSubscriber())
             PollMessages.setVote(adminId, messageId, option, vote = true)
             awaitBrokering()
-            val actual = subscriber.values().map { (it as UpdatedMessage).getMessageId() }
+            val actual = subscriber.values().map { (it as UpdatedPollMessage).getMessageId() }
             assertEquals(listOf(messageId), actual)
         }
     }